+++ src/compas/geometry/primitives/circle.py
@@ -1,3 +1,9 @@
+"""
+.. testsetup::
+
+    from compas.geometry import Circle
+
+"""
 from __future__ import print_function
 from __future__ import absolute_import
 from __future__ import division
@@ -54,6 +60,42 @@ class Circle(Primitive):
         self._radius = None
         self.plane = plane
         self.radius = radius
+
+    @property
+    def DATASCHEMA(self):
+        import schema
+        from compas.data import is_float3
+        return schema.Schema({
+            "plane": schema.And(
+                lambda x: is_float3(x[0]),
+                lambda x: is_float3(x[1])
+            ),
+            "radius": schema.And(float, lambda x: x > 0)
+        })
+
+    @property
+    def JSONSCHEMA(self):
+        from compas import versionstring
+        schema = {
+            "$schema": "http://json-schema.org/draft-07/schema#",
+            "$id": "https://github.com/compas-dev/compas/schemas/circle.json",
+            "$compas": versionstring,
+            "type": "object",
+            "properties": {
+                "plane": {
+                    "type": "array",
+                    "minItems": 2,
+                    "maxItems": 2,
+                    "items": [
+                        {"type": "array", "minItems": 3, "maxItems": 3, "items": {"type": "number"}},
+                        {"type": "array", "minItems": 3, "maxItems": 3, "items": {"type": "number"}}
+                    ]
+                },
+                "radius": {"type": "number", "exclusiveMinimum": 0}
+            },
+            "required": ["plane", "radius"]
+        }
+        return schema
 
     @property
     def data(self):
@@ -117,7 +159,7 @@ class Circle(Primitive):
     # ==========================================================================
 
     def __repr__(self):
-        return 'Circle({0}, {1})'.format(self.plane, self.radius)
+        return 'Circle({0!r}, {1!r})'.format(self.plane, self.radius)
 
     def __len__(self):
         return 2
@@ -199,12 +241,3 @@ class Circle(Primitive):
         >>> circle.transform(T)
         """
         self.plane.transform(T)
-
-
-# ==============================================================================
-# Main
-# ==============================================================================
-
-if __name__ == "__main__":
-    import doctest
-    doctest.testmod(globs=globals())