+++ src/compas/geometry/primitives/plane.py
@@ -1,3 +1,9 @@
+"""
+.. testsetup::
+
+    from compas.geometry import Plane
+
+"""
 from __future__ import print_function
 from __future__ import absolute_import
 from __future__ import division
@@ -51,6 +57,31 @@ class Plane(Primitive):
         self.normal = normal
 
     @property
+    def DATASCHEMA(self):
+        from schema import Schema
+        from compas.data import is_float3
+        return Schema({
+            "point": is_float3,
+            "normal": is_float3
+        })
+
+    @property
+    def JSONSCHEMA(self):
+        from compas import versionstring
+        schema = {
+            "$schema": "http://json-schema.org/draft-07/schema#",
+            "$id": "https://github.com/compas-dev/compas/schemas/plane.json",
+            "$compas": versionstring,
+            "type": "object",
+            "properties": {
+                "point": {"type": "array", "minItems": 3, "maxItems": 3, "items": {"type": "number"}},
+                "normal": {"type": "array", "minItems": 3, "maxItems": 3, "items": {"type": "number"}}
+            },
+            "required": ["point", "normal"]
+        }
+        return schema
+
+    @property
     def data(self):
         """dict : The data dictionary that represents the plane."""
         return {'point': list(self.point),
@@ -99,7 +130,7 @@ class Plane(Primitive):
     # ==========================================================================
 
     def __repr__(self):
-        return 'Plane({0}, {1})'.format(self.point, self.normal)
+        return 'Plane({0!r}, {1!r})'.format(self.point, self.normal)
 
     def __len__(self):
         return 2
@@ -258,6 +289,7 @@ class Plane(Primitive):
 
         Examples
         --------
+        >>> from compas.geometry import Frame
         >>> frame = Frame([1, 1, 1], [0.68, 0.68, 0.27], [-0.67, 0.73, -0.15])
         >>> Plane.from_frame(frame)
         Plane(Point(1.000, 1.000, 1.000), Vector(-0.299, -0.079, 0.951))
@@ -288,14 +320,3 @@ class Plane(Primitive):
         """
         self.point.transform(T)
         self.normal.transform(T)
-
-
-# ==============================================================================
-# Main
-# ==============================================================================
-
-if __name__ == '__main__':
-
-    import doctest
-    from compas.geometry import Frame  # noqa F401
-    doctest.testmod(globs=globals())