+++ src/compas/geometry/primitives/point.py
@@ -1,3 +1,9 @@
+"""
+.. testsetup::
+
+    from compas.geometry import Point
+
+"""
 from __future__ import print_function
 from __future__ import absolute_import
 from __future__ import division
@@ -120,6 +126,26 @@ class Point(Primitive):
         self.x = x
         self.y = y
         self.z = z
+
+    @property
+    def DATASCHEMA(self):
+        from schema import Schema
+        from compas.data import is_float3
+        return Schema(is_float3)
+
+    @property
+    def JSONSCHEMA(self):
+        from compas import versionstring
+        schema = {
+            "$schema": "http://json-schema.org/draft-07/schema#",
+            "$id": "https://github.com/compas-dev/compas/schemas/point.json",
+            "$compas": versionstring,
+            "type": "array",
+            "minItems": 3,
+            "maxItems": 3,
+            "items": {"type": "number"}
+        }
+        return schema
 
     @property
     def data(self):
@@ -408,7 +434,8 @@ class Point(Primitive):
 
         Examples
         --------
-        >>> T = Translation([1.0, 2.0, 3.0])
+        >>> from compas.geometry import Translation
+        >>> T = Translation.from_vector([1.0, 2.0, 3.0])
         >>> a = Point(0.0, 0.0, 0.0)
         >>> points = [a]
         >>> Point.transform_collection(points, T)
@@ -440,7 +467,8 @@ class Point(Primitive):
 
         Examples
         --------
-        >>> T = Translation([1.0, 2.0, 3.0])
+        >>> from compas.geometry import Translation
+        >>> T = Translation.from_vector([1.0, 2.0, 3.0])
         >>> a = Point(0.0, 0.0, 0.0)
         >>> points = [a]
         >>> points = Point.transformed_collection(points, T)
@@ -741,7 +769,7 @@ class Point(Primitive):
         --------
         >>> from compas.geometry import Translation
         >>> point = Point(0.0, 0.0, 0.0)
-        >>> T = Translation([1.0, 1.0, 1.0])
+        >>> T = Translation.from_vector([1.0, 1.0, 1.0])
         >>> point.transform(T)
         >>> point.x == 1.0
         True
@@ -750,15 +778,3 @@ class Point(Primitive):
         self.x = point[0]
         self.y = point[1]
         self.z = point[2]
-
-
-# ==============================================================================
-# Main
-# ==============================================================================
-
-if __name__ == '__main__':
-
-    import doctest
-    from compas.geometry import Translation  # noqa F401
-
-    doctest.testmod(globs=globals())