+++ src/compas/geometry/primitives/quaternion.py
@@ -1,3 +1,9 @@
+"""
+.. testsetup::
+
+    from compas.geometry import Quaternion
+
+"""
 from __future__ import print_function
 from __future__ import absolute_import
 from __future__ import division
@@ -15,7 +21,7 @@ from compas.geometry import quaternion_from_matrix
 from compas.geometry.primitives import Primitive
 
 
-__all__ = ['Quaternion']
+__all__ = ["Quaternion"]
 
 
 class Quaternion(Primitive):
@@ -104,7 +110,7 @@ class Quaternion(Primitive):
     True
     """
 
-    __slots__ = ['_w', '_x', '_y', '_z']
+    __slots__ = ["_w", "_x", "_y", "_z"]
 
     def __init__(self, w, x, y, z):
         super(Quaternion, self).__init__()
@@ -116,6 +122,40 @@ class Quaternion(Primitive):
         self.x = x
         self.y = y
         self.z = z
+
+    @property
+    def DATASCHEMA(self):
+        from schema import Schema
+        return Schema({"w": float, "x": float, "y": float, "z": float})
+
+    @property
+    def JSONSCHEMA(self):
+        from compas import versionstring
+        schema = {
+            "$schema": "http://json-schema.org/draft-07/schema#",
+            "$id": "https://github.com/compas-dev/compas/schemas/quaternion.json",
+            "$compas": versionstring,
+            "type": "object",
+            "properties": {
+                "w": {"type": "number"},
+                "x": {"type": "number"},
+                "y": {"type": "number"},
+                "z": {"type": "number"}
+            },
+            "required": ["w", "x", "y", "z"]
+        }
+        return schema
+
+    @property
+    def data(self):
+        return {"w": self.w, "x": self.x, "y": self.y, "z": self.z}
+
+    @data.setter
+    def data(self, data):
+        self.w = data["w"]
+        self.x = data["x"]
+        self.y = data["y"]
+        self.z = data["z"]
 
     @property
     def w(self):
@@ -152,17 +192,6 @@ class Quaternion(Primitive):
     @z.setter
     def z(self, z):
         self._z = float(z)
-
-    @property
-    def data(self):
-        return {'w': self.w, 'x': self.x, 'y': self.y, 'z': self.z}
-
-    @data.setter
-    def data(self, data):
-        self.w = data['w']
-        self.x = data['x']
-        self.y = data['y']
-        self.z = data['z']
 
     @property
     def wxyz(self):
@@ -217,7 +246,7 @@ class Quaternion(Primitive):
         raise KeyError
 
     def __eq__(self, other, tol=1e-05):
-        if not hasattr(other, '__iter__') or not hasattr(other, '__len__') or len(self) != len(other):
+        if not hasattr(other, "__iter__") or not hasattr(other, "__len__") or len(self) != len(other):
             return False
         for v1, v2 in zip(self, other):
             if math.fabs(v1 - v2) > tol:
@@ -231,7 +260,7 @@ class Quaternion(Primitive):
         return 4
 
     def __repr__(self):
-        return 'Quaternion({:.{prec}f}, {:.{prec}f}, {:.{prec}f}, {:.{prec}f})'.format(self.w, self.x, self.y, self.z, prec=3)
+        return "Quaternion({:.{prec}f}, {:.{prec}f}, {:.{prec}f}, {:.{prec}f})".format(self.w, self.x, self.y, self.z, prec=3)
 
     def __mul__(self, other):
         """Multiply operator for two quaternions.
@@ -286,7 +315,7 @@ class Quaternion(Primitive):
         --------
         >>>
         """
-        return cls(data['w'], data['x'], data['y'], data['z'])
+        return cls(data["w"], data["x"], data["y"], data["z"])
 
     @classmethod
     def from_frame(cls, frame):
@@ -303,11 +332,12 @@ class Quaternion(Primitive):
 
         Examples
         --------
+        >>> from compas.geometry import allclose
         >>> from compas.geometry import Frame
         >>> q = [1., -2., 3., -4.]
         >>> F = Frame.from_quaternion(q)
         >>> Q = Quaternion.from_frame(F)
-        >>> allclose(list(Q.canonized()), quaternion_canonize(quaternion_unitize(q)))
+        >>> allclose(Q.canonized(), quaternion_canonize(quaternion_unitize(q)))
         True
         """
         w, x, y, z = frame.quaternion
@@ -462,16 +492,3 @@ class Quaternion(Primitive):
         """
         qc = quaternion_conjugate(self)
         return Quaternion(*qc)
-
-
-# ==============================================================================
-# Main
-# ==============================================================================
-
-if __name__ == "__main__":
-
-    import doctest
-
-    from compas.geometry import allclose  # noqa: F401
-
-    doctest.testmod(globs=globals())