+++ src/compas/geometry/transformations/reflection.py
@@ -40,6 +40,9 @@ class Reflection(Transformation):
             pass
         super(Reflection, self).__init__(matrix=matrix)
 
+    def __repr__(self):
+        return "Reflection({0!r})".format(self.matrix)
+
     @classmethod
     def from_plane(cls, plane):
         """Creates a reflection object that mirrors wrt the given plane.
@@ -86,13 +89,3 @@ class Reflection(Transformation):
             point = frame.point
             zaxis = frame.zaxis
         return cls.from_plane((point, zaxis))
-
-
-# ==============================================================================
-# Main
-# ==============================================================================
-
-if __name__ == "__main__":
-
-    import doctest
-    doctest.testmod(globs=globals())