+++ src/compas/geometry/transformations/shear.py
@@ -54,6 +54,9 @@ class Shear(Transformation):
             pass
         super(Shear, self).__init__(matrix=matrix)
 
+    def __repr__(self):
+        return "Shear({0!r})".format(self.matrix)
+
     @classmethod
     def from_angle_direction_plane(cls, angle, direction, plane):
         """
@@ -79,6 +82,7 @@ class Shear(Transformation):
 
         Examples
         --------
+        >>> from compas.geometry import cross_vectors
         >>> angle = 0.1
         >>> direction = [0.1, 0.2, 0.3]
         >>> point = [4, 3, 1]
@@ -102,17 +106,3 @@ class Shear(Transformation):
         >>> S = Shear.from_entries([1, 2, 3])
         """
         return cls(matrix_from_shear_entries(shear_entries))
-
-
-# ==============================================================================
-# Main
-# ==============================================================================
-
-if __name__ == "__main__":
-
-    import doctest
-
-    from compas.geometry import Shear  # noqa: F401 F811
-    from compas.geometry import cross_vectors  # noqa: F401
-
-    doctest.testmod(globs=globals())