+++ src/compas/topology/__init__.py
@@ -36,8 +36,10 @@ Orientation
 
     face_adjacency
     face_adjacency_numpy
+    face_adjacency_rhino
     unify_cycles
     unify_cycles_numpy
+    unify_cycles_rhino
 
 
 Traversal
@@ -47,6 +49,7 @@ Traversal
     :toctree: generated/
     :nosignatures:
 
+    astar_lightest_path
     astar_shortest_path
     breadth_first_ordering
     breadth_first_traverse
@@ -67,6 +70,7 @@ from .traversal import (
     breadth_first_traverse,
     breadth_first_paths,
     shortest_path,
+    astar_lightest_path,
     astar_shortest_path,
     dijkstra_distances,
     dijkstra_path
@@ -99,6 +103,7 @@ __all__ = [
     'breadth_first_traverse',
     'breadth_first_paths',
     'shortest_path',
+    'astar_lightest_path',
     'astar_shortest_path',
     'dijkstra_distances',
     'dijkstra_path',